test transactionslog
I added some additional code to my initial transaction.cpp file to test the functionality of the code to ensure it worked. This is  the output file of what it does. I may edit the transaction.cpp file to add more details such as failed withdraws or something later, but this should satisfy the requirements for now